Use consistent usize index type for all child iterators
BooleanOpChildIter, FloatOpChildIter, and StringOpChildIter used u8 for
their index field, while BitVecOpChildIter already used usize. Using u8
limits n-ary operations (like And/Or) to 255 children and is inconsistent.

Change the index field to usize in all three iterators and remove the
now-unnecessary `as usize` casts.

+++ crates/clarirs_core/src/ast/bool.rs
@@ -45,7 +45,7 @@ pub type BoolAst<'c> = AstRef<'c, BooleanOp<'c>>;
 
 pub struct BooleanOpChildIter<'a, 'c> {
     op: &'a BooleanOp<'c>,
-    index: u8,
+    index: usize,
 }
 
 impl<'c> BooleanOp<'c> {
@@ -131,11 +131,11 @@ impl<'a, 'c> Iterator for BooleanOpChildIter<'a, 'c> {
             (BooleanOp::ITE(_, _, c), 2) => Some(c.into()),
 
             // N-ary variants (And/Or with Vec)
-            (BooleanOp::And(args), i) if (i as usize) < args.len() => {
-                Some((&args[i as usize]).into())
+            (BooleanOp::And(args), i) if i < args.len() => {
+                Some((&args[i]).into())
             }
-            (BooleanOp::Or(args), i) if (i as usize) < args.len() => {
-                Some((&args[i as usize]).into())
+            (BooleanOp::Or(args), i) if i < args.len() => {
+                Some((&args[i]).into())
             }
 
             _ => None,
@@ -194,7 +194,7 @@ impl<'a, 'c> ExactSizeIterator for BooleanOpChildIter<'a, 'c> {
             BooleanOp::And(args) => args.len(),
             BooleanOp::Or(args) => args.len(),
         };
-        total.saturating_sub(self.index as usize)
+        total.saturating_sub(self.index)
     }
 }
